+++ src/main/java/com/dataracy/modules/reference/application/service/query/AnalysisPurposeQueryService.java
@@ -3,8 +3,8 @@ package com.dataracy.modules.reference.application.service.query;
 import com.dataracy.modules.reference.application.dto.response.allview.AllAnalysisPurposesResponse;
 import com.dataracy.modules.reference.application.dto.response.singleview.AnalysisPurposeResponse;
 import com.dataracy.modules.reference.application.mapper.AnalysisPurposeDtoMapper;
-import com.dataracy.modules.reference.application.port.in.analysis_purpose.FindAllAnalysisPurposesUseCase;
-import com.dataracy.modules.reference.application.port.in.analysis_purpose.FindAnalysisPurposeUseCase;
+import com.dataracy.modules.reference.application.port.in.analysispurpose.FindAllAnalysisPurposesUseCase;
+import com.dataracy.modules.reference.application.port.in.analysispurpose.FindAnalysisPurposeUseCase;
 import com.dataracy.modules.reference.application.port.out.AnalysisPurposeRepositoryPort;
 import com.dataracy.modules.reference.domain.exception.ReferenceException;
 import com.dataracy.modules.reference.domain.model.AnalysisPurpose;
