+++ src/main/java/com/dataracy/modules/reference/application/service/query/DataSourceQueryService.java
@@ -3,8 +3,8 @@ package com.dataracy.modules.reference.application.service.query;
 import com.dataracy.modules.reference.application.dto.response.allview.AllDataSourcesResponse;
 import com.dataracy.modules.reference.application.dto.response.singleview.DataSourceResponse;
 import com.dataracy.modules.reference.application.mapper.DataSourceDtoMapper;
-import com.dataracy.modules.reference.application.port.in.data_source.FindAllDataSourcesUseCase;
-import com.dataracy.modules.reference.application.port.in.data_source.FindDataSourceUseCase;
+import com.dataracy.modules.reference.application.port.in.datasource.FindAllDataSourcesUseCase;
+import com.dataracy.modules.reference.application.port.in.datasource.FindDataSourceUseCase;
 import com.dataracy.modules.reference.application.port.out.DataSourceRepositoryPort;
 import com.dataracy.modules.reference.domain.exception.ReferenceException;
 import com.dataracy.modules.reference.domain.model.DataSource;
