+++ src/main/java/com/dataracy/modules/user/application/service/command/UserCommandService.java
@@ -5,10 +5,10 @@ import com.dataracy.modules.auth.application.port.in.jwt.JwtGenerateUseCase;
 import com.dataracy.modules.auth.application.port.in.jwt.JwtValidateUseCase;
 import com.dataracy.modules.auth.application.port.in.redis.TokenRedisUseCase;
 import com.dataracy.modules.common.support.lock.DistributedLock;
-import com.dataracy.modules.reference.application.port.in.author_level.FindAuthorLevelUseCase;
+import com.dataracy.modules.reference.application.port.in.authorlevel.FindAuthorLevelUseCase;
 import com.dataracy.modules.reference.application.port.in.occupation.FindOccupationUseCase;
 import com.dataracy.modules.reference.application.port.in.topic.IsExistTopicUseCase;
-import com.dataracy.modules.reference.application.port.in.visit_source.FindVisitSourceUseCase;
+import com.dataracy.modules.reference.application.port.in.visitsource.FindVisitSourceUseCase;
 import com.dataracy.modules.user.application.dto.request.ChangePasswordRequest;
 import com.dataracy.modules.user.application.dto.request.OnboardingRequest;
 import com.dataracy.modules.user.application.dto.request.SelfSignUpRequest;
